+++ CHANGELOG.md
@@ -1,5 +1,32 @@
-## Changelog
-### Version 1.3 (Dec 22, 2020)
+# Changelog
+
+### Version 1.5 (July 19, 2022)
+### New features and improvements
+***
+* Add support for vpc endpoint
+
+### Dependency updates
+***
+* alibabacloud-credentials from 1.0 to 1.1
+
+## Version 1.5 (July 19, 2022)
+### New features and improvements
+***
+* Add noun notes
+
+### Bug fixes
+***
+* configuration cluster save updated System Disk Category and System Disk size
+
+## Version 1.4 (July 18, 2022)
+### New features and improvements
+*** 
+* Add support system disk category
+* Add support system disk size
+* Add attach Public Ip
+
+
+## Version 1.3 (Dec 22, 2020)
 - Add chinese readme index.
 - Delete allocate public ip method.